+++ src/main/java/domain/CardMachine.java
@@ -1,7 +1,6 @@
 package domain;
 
 import constant.Rank;
-import constant.Suit;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Collections;
